# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "After-Sales Work Order Review API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/workorders": {
      "get": {
        "tags": [
          "review"
        ],
        "summary": "List Workorders",
        "description": "工单列表（分页 + 搜索）。\n\n- status: 按工单状态筛选（pending_review / confirmed / stashed）\n- keyword: 按序列号/站点/客户/项目名模糊搜索\n- offset/limit: 分页参数（默认每页 50 条）",
        "operationId": "list_workorders_api_workorders_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "keyword",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Keyword"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedWorkOrderSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/workorders/{workorder_id}": {
      "get": {
        "tags": [
          "review"
        ],
        "summary": "Get Workorder",
        "operationId": "get_workorder_api_workorders__workorder_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "workorder_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workorder Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkOrderResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/workorders/{workorder_id}/review": {
      "post": {
        "tags": [
          "review"
        ],
        "summary": "Review Workorder",
        "description": "[已废弃] 请使用 POST /{workorder_id}/confirm 代替。\n\n本端点仍可正常使用，但将在未来版本中移除。",
        "operationId": "review_workorder_api_workorders__workorder_id__review_post",
        "deprecated": true,
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "workorder_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workorder Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/workorders/{workorder_id}/confirm": {
      "post": {
        "tags": [
          "review"
        ],
        "summary": "Confirm Workorder",
        "description": "确认提交：本地落库后立即返回，后台异步同步至销售易。",
        "operationId": "confirm_workorder_api_workorders__workorder_id__confirm_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "workorder_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workorder Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfirmResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/workorders/{workorder_id}/stash": {
      "post": {
        "tags": [
          "review"
        ],
        "summary": "Stash Workorder",
        "description": "暂存审核进度。\n\nmode='manual': 标记工单为 stashed + 释放锁，其他人可接手。\nmode='auto_save': 仅保存进度，不改变工单状态，不释放锁。",
        "operationId": "stash_workorder_api_workorders__workorder_id__stash_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "workorder_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workorder Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StashRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StashResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "review"
        ],
        "summary": "Get Stash",
        "description": "获取暂存的审核进度。无暂存时返回空数据，避免将正常初始状态报为 404。",
        "operationId": "get_stash_api_workorders__workorder_id__stash_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "workorder_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workorder Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StashData"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "review"
        ],
        "summary": "Delete Stash",
        "description": "删除暂存的审核进度（用户丢弃修改时调用）。",
        "operationId": "delete_stash_api_workorders__workorder_id__stash_delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "workorder_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workorder Id"
            }
          },
          {
            "name": "X-Lock-Fencing-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "X-Lock-Fencing-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StashResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/workorders/{workorder_id}/audit-logs": {
      "get": {
        "tags": [
          "review"
        ],
        "summary": "Get Audit Logs",
        "operationId": "get_audit_logs_api_workorders__workorder_id__audit_logs_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "workorder_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workorder Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AuditLogEntry"
                  },
                  "title": "Response Get Audit Logs Api Workorders  Workorder Id  Audit Logs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/import": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Import New Workorders",
        "description": "将源表新增工单幂等导入审核队列。",
        "operationId": "import_new_workorders_api_admin_import_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/admin/sync-failures": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "List Sync Failures",
        "description": "列出同步失败的工单（sync_status='failed'）。",
        "operationId": "list_sync_failures_api_admin_sync_failures_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/sync-failures/{workorder_id}/retry": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Retry Sync",
        "description": "手动重试指定工单的销售易同步。\n\n支持 failed / pending 状态的工单重试。已存在 sync_external_id\n的记录（幂等已完成）将被直接标记为 synced 而不重复调用 API。",
        "operationId": "retry_sync_api_admin_sync_failures__workorder_id__retry_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "workorder_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workorder Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/sync-uncertain/{workorder_id}/reconcile": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Reconcile Uncertain Sync",
        "description": "人工确认销售易已创建工单，并绑定其外部工单号。",
        "operationId": "reconcile_uncertain_sync_api_admin_sync_uncertain__workorder_id__reconcile_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "workorder_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workorder Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReconcileSyncRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/sync-uncertain/{workorder_id}/confirm-not-created": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Confirm Uncertain Not Created",
        "description": "人工确认销售易未创建工单，将 uncertain 转为可人工重试的 failed。",
        "operationId": "confirm_uncertain_not_created_api_admin_sync_uncertain__workorder_id__confirm_not_created_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "workorder_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workorder Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PaginatedWorkOrderSummary": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/WorkOrderSummary"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "offset",
          "limit"
        ],
        "title": "PaginatedWorkOrderSummary",
        "description": "分页工单摘要列表。"
      },
...
      "ReviewRequest": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "changes": {
            "items": {
              "$ref": "#/components/schemas/FieldChange"
            },
            "type": "array",
            "title": "Changes"
          },
          "reject_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reject Reason"
          },
          "review_notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Review Notes"
          },
          "lock_fencing_token": {
            "type": "integer",
            "title": "Lock Fencing Token"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "version",
          "lock_fencing_token"
        ],
        "title": "ReviewRequest"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}